Move Course test instances into separate module
